+++ renderer/src/App.tsx
@@ -1,5 +1,6 @@
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom'
 import './App.css'
+import Welcome from './pages/Welcome'
 import Onboarding from './pages/Onboarding'
 import WalletConfig from './pages/WalletConfig'
 import Dashboard from './pages/Dashboard'
@@ -17,12 +18,13 @@ const App = ():JSX.Element => {
       <Saturn />
       <Router>
         <Routes>
-          <Route path="/" element={<Onboarding />} />
+          <Route path="/" element={<Welcome />} />
           <Route path="*" element={
             <>
               <Sentry />
               <Plausible />
               <Routes>
+                <Route path="/onboarding" element={<Onboarding />} />
                 <Route path="/wallet" element={<WalletConfig />} />
                 <Route path="/dashboard" element={<Dashboard />} />
               </Routes>
